atualizado

+++ src/controllers/brandsController.js
@@ -1,25 +1,32 @@
+const { connect } = require('../data/database')
 const mongoose = require('mongoose')
 const Brand = require('../models/brandsModels')
 
+connect()
+
+const jwt = require('jsonwebtoken')
+const SECRET = process.env.SECRET
+
 //Mostrar todos as marcas
 const getAll = async (req, res) => {
-    try {
-        const list = await Brand.find()
-        res.json(list)
-    } catch (err) {
-        res.status(500).json({message: err.message})
-    }
-}
+
+        try {
+            const brands = await Brand.find()
+            res.json(brands)
+        } catch (err) {
+            res.status(500).json({message: err.message})
+        }
+      }
 
 //Mostrar uma
 const getOne = async (req, res) => {
-    res.send(res.list)
+    res.send(res.brands)
 
 }
 
 //Criar uma
 const createOne = async (req, res) => {
-    const list = new Brand({
+    const brands = new Brand({
         _id: new mongoose.Types.ObjectId(),
         marca: req.body.marca,
         empresa: req.body.empresa,
@@ -28,7 +35,7 @@ const createOne = async (req, res) => {
     })
 
     try {
-        const newList = await list.save()
+        const newList = await brands.save()
         res.status(201).json(newList)
     } catch (err) {
         res.status(400).json({ message: err.message})
@@ -38,11 +45,11 @@ const createOne = async (req, res) => {
 //Atualizar uma
 const updateOne = async (req, res) => {
     if (req.body.marca != null) {
-        res.list.marca = req.body.marca
+        res.brands.marca = req.body.marca
     }
 
     try {
-        const updateList = await res.list.save()
+        const updateList = await res.brands.save()
         res.json(updateList)
     } catch (err) {
         res.status(400).json({ message: err.message})
@@ -50,10 +57,10 @@ const updateOne = async (req, res) => {
 }
 
 
-//Deletar um
+//Deletar uma
 const deleteOne = async (req, res) => {
     try {
-        await res.list.remove()
+        await res.brands.remove()
         res.json({message: 'Marca deletada!'})
     } catch (err) {
         res.status(500).json({ message: err.message})
@@ -62,10 +69,10 @@ const deleteOne = async (req, res) => {
 
 //função "marca nao encontrada"
 async function getList(req,res,next){
-    let list
+    let brands
     try {
-        list = await Brand.findById(req.params.id)
-        if (list == null) {
+        brands = await Brand.findById(req.params.id)
+        if (brands == null) {
             return res.status(404).json({message: 'Marca não encontrada'})
         }
 
@@ -74,9 +81,27 @@ async function getList(req,res,next){
 
     }
 
-    res.list = list
+    res.brands = brands
     next()
 
 }
 
-module.exports = { getAll, getOne, createOne, updateOne, deleteOne, getList }
+async function auth (req, res) {
+    const authHeader = req.get("Autorização")
+
+    if (!authHeader) {
+      return res.status(401).send("Precisa de um login!")
+    }
+
+    const token = authHeader.split(" ")[1]
+
+    jwt.verify(token, SECRET, (err) => {
+      if (err) {
+        return res.status(403).send("Invalid token!")
+      }
+
+      return token
+    })
+  }
+
+module.exports = { getAll, getOne, createOne, updateOne, deleteOne, getList, auth }
